all functions compiling

+++ Libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: cghanime <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/09 20:09:18 by cghanime          #+#    #+#             */
-/*   Updated: 2018/11/19 16:41:55 by cghanime         ###   ########.fr       */
+/*   Updated: 2018/11/20 13:07:32 by cghanime         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(const char *s)
+size_t		ft_strlen(const char *s)
 {
 	size_t i;
